cosmetics, multiline string syntax
