+++ GraphQL-Core/Extensions/GraphQLExtensions.cs
@@ -21,7 +21,6 @@ namespace GraphQLCore.Extensions
                 bool nullable = false;
                 Type baseGraphQLType = null;
 
-
                 if (cSharpType.IsEnum)
                 {
                     returnType = typeof(EnumerationGraphType<>).MakeGenericType(cSharpType);
@@ -117,6 +116,7 @@ namespace GraphQLCore.Extensions
                             {
                                 var genericType = typeof(GenericType<>).MakeGenericType(str);
                                 var derived = genericType.GetDerivedGenericUserType();
+
                                 returnType = derived;
                                 baseGraphQLType = returnType;
                                 nullable = true;
@@ -126,7 +126,6 @@ namespace GraphQLCore.Extensions
                             break;
                     }
                 }
-
                 else if (cSharpType.IsClass && (
                     false
                     || cSharpType.Implements(typeof(IList))
@@ -143,8 +142,10 @@ namespace GraphQLCore.Extensions
                         return (null, false, null);
 
                     (var underClass, _, _) = GetConvertibleBaseCSharpType(listType);
+
                     if (!underClass.Implements(typeof(IGraphType)))
                         underClass = underClass.GetGraphTypeFromType(true);
+
                     returnType = typeof(ListGraphType<>).MakeGenericType(underClass);
                     baseGraphQLType = underClass;
                     nullable = true;
@@ -161,7 +162,7 @@ namespace GraphQLCore.Extensions
 
                 return (returnType, nullable, baseGraphQLType);
             }
-            catch(Exception e)
+            catch (Exception e)
             {
                 throw new Exception("Attempt to convert the type {cSharpType.Name} to it's GraphQL-equivalent failed", e);
             }
@@ -174,7 +175,7 @@ namespace GraphQLCore.Extensions
                 var type = cSharpType.ConvertToGraphQLType();
                 return type;
             }
-            catch(Exception e)
+            catch (Exception e)
             {
                 return null;
             }
@@ -189,10 +190,10 @@ namespace GraphQLCore.Extensions
         {
             (Type graphQlType, var nullable, var underClass) = GetConvertibleBaseCSharpType(cSharpType);
 
-            if(graphQlType is null)
+            if (graphQlType is null)
                 throw new GraphQLCoreConversionException(
-                    $"The C# type '{cSharpType.Name}' is not currently supported for auto-conversion into a GraphQL Node-type. " +
-                    $"Consider registering it as a custom class in your GraphQL middleware");
+                    $@"The C# type '{cSharpType.Name}' is not currently supported for auto-conversion into a GraphQL Node-type.
+                    Consider registering it as a custom class in your GraphQL middleware");
 
             if (graphQlType.Implements(typeof(IGraphType)))
                 return graphQlType;
@@ -213,8 +214,8 @@ namespace GraphQLCore.Extensions
 
             if (graphQlType is null)
                 throw new GraphQLCoreConversionException(
-                    $"The C# type '{cSharpType.Name}' is not currently supported for auto-conversion into a GraphQL Node-type. " +
-                    $"Consider registering it as a custom class in your GraphQL middleware");
+                    $@"The C# type '{cSharpType.Name}' is not currently supported for auto-conversion into a GraphQL Node-type.
+                    Consider registering it as a custom class in your GraphQL middleware");
 
             if (underClass.Implements(typeof(IGraphType)))
                 return underClass;
