+++ GraphQL-Core/Extensions/GraphQLNETValueConversions.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace GraphQLCore.Extensions
 {
